backend/requirements.txt
anyio==3.7.1
asgiref==3.9.1
certifi==2025.8.3
click==8.2.1
distro==1.9.0
fastapi==0.68.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
lxml==6.0.1
openai==1.107.1
pydantic==1.10.22
PyPDF2==3.0.1
python-docx==0.8.11
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.1
starlette==0.14.2
tqdm==4.67.1
typing_extensions==4.15.0
uvicorn==0.15.0
